backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
fastapi-cors
pandas
openpyxl
psycopg2-binary
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
alembic
